Initial draft
We implement fa_link() which is supposed to be called by any component
which wants to use Font Awesome, and includes a minimally modified
font-awesome.css at most *once*.

The modifications to font-awesome.css are necessary so Font Awesome can
also be used for secondary languages. Since the base path isn't
identical for the primary language and secondary languages, we generate
the important @font-face at-rule dynamically.