Close #13: fontawesome.min.css not found
We remove the ugly workaround of trying to include fontawesome.min.css
twice, so that one of the relative URLs will work regardless of whether
the main or a secondary language is requested, but the other would
fail. Instead we're requiring an absolute URL, which doesn't appear to
actually be a problem, as commit f1eed5f requires that for the
external_plugins configuration, anyway.