Fix #20: block_formats dropdown may malfunction
We remove the %PAGEHEADERS% placeholder from the init to avoid trouble
with CMSimple_XH 1.7. For CMSimple_XH 1.6 this change doesn't allow to
choose a page splitting heading from the block_formats dropdown, but
that might be best for many users, anyway, and advanced users are still
able to insert the %PAGEHEADERS% placeholder as needed.