Close #23: Toolbar icon is confusing
We choose the fort-awesome icon for both editors, since we have already
chosen it as plugin logo and it is supposed to be much more unique than
the font-awesome icon. We use a PNG for simplicity.

We also fix the font awesome button tooltip of the TinyMCE 4 plugin,
and move the button down beside the charmap button, were it appears to
be much more appropriate.