Close #38: Optional support for Font Awesome 5
We add the relevant CSS and font files from Font Awesome 5.15.2, and
allow users to choose between version 4 and 5.  For Font Awesome 5, we
also give option to use the version 4 shims.

Unfortunately, the editor plugins do not yet support Font Awesome 5,
so we stick with Font Awesome 4 as default, what is best for backward
compatibility anyway.

According to #3, we don't include the SVG fonts.

Co-authored-by: Christoph M. Becker <cmbecker69@gmx.de>