Extract icon metadata
The icon metadata are different for FA versions, so we extract them
into a separate script, which we're loading dynamically.  This relies
currently on the script being loaded before a user opens the dialog,
which is likely, but not certain.  We should cater to that, but for now
the solution is sufficient.