Support FontAwesome 5
Since the metadata format has changed, we need to cater to that in
gen_icon_data.php.  One particular issues is that not all FA 5 icons
have categories, so we would either need to come up with a category
for those (say, "Miscelleaneous"), or we add a category "All" for all
icons – we choose the latter for now.

We need to make minor adjustments to the actual editor plugin, and also
introduce a TinyMCE configuration parameter ("fontawesome_version"), so
users can actually use FontAwesome 5.