Dynamically link to proper Font Awesome stylesheet
Since we're using classic mode, the editor instance is in an iframe,
so we need to link to the Font Awesome stylesheet to see the icons in
the editor.  This is currently hard-coded in the editor inits, but
that doesn't cater to other Font Awesome versions than 4.  To avoid
users to configure that manually, we link to the proper stylesheet
dynamically.

To get the URL also right for secondary languages, but to avoid some
additional meta information somewhere, we figure out the base path
from the common prefix of the document's base URI and the plugin URL.