Load shims in editor if configured
The editor preview should be as close to the actual view, so we need to
load the shims, if these are configured to be loaded.

This is typically important for brand icons which have been inserted
with previous editor plugin versions, which won't show up without the
shims at all.